old-tests/submission/Infraware/Forms/contents/Forms/Not_Use/input_required.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>input_required</h3>
	</p>
	
	<li>When the inputbox control is empty, you cannot submit this form.</li>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="?x=1"
		  name="input_form">
	<p><input type='text' id='input_text'></p>
	<p><input type='submit' id='input_submit'></p>
	</form>

	<script>
		
		var input_text = document.getElementById("input_text");
		input_text.required = true;

		if (typeof(input_text.required) == "boolean") {
			if (input_text.required != true) {
				test(function() {
					assert_true(false, "required attribute is not correct.");
				});
			}
		} else {
			test(function() {
				assert_unreached("required attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
